get examples and tests compiling
