+++ examples/v4/presale/start-with-allowlist.ts
@@ -71,6 +71,7 @@ async function startAllowlistedPresaleExample() {
         image: 'ipfs://bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi',
         tokenAdmin: account.address,
         metadata: { description: 'A token created through allowlisted presale' },
+        // biome-ignore lint: TODO come back to type these
         presale: { bps: presaleConfig.presaleSupplyBps! },
       },
       presaleConfig,
