--- test/config/clankerTokenV4.presale-converter.test.ts
@@ -1,230 +1,0 @@
-import { describe, expect, test } from 'bun:test';
-import type { ClankerTokenV4 } from '../../src/config/clankerTokenV4.js';
-import { clankerTokenV4PresaleConverter } from '../../src/config/clankerTokenV4.js';
-
-describe('clankerTokenV4PresaleConverter', () => {
-  const baseConfig: ClankerTokenV4 = {
-    name: 'Test Presale Token',
-    symbol: 'TPT',
-    tokenAdmin: '0x5b32C7635AFe825703dbd446E0b402B8a67a7051',
-    chainId: 8453,
-    presale: {
-      minEthGoal: 1,
-      maxEthGoal: 10,
-      presaleDuration: 3600,
-      recipient: '0x5b32C7635AFe825703dbd446E0b402B8a67a7051',
-      lockupDuration: 86400,
-      vestingDuration: 86400,
-    },
-  };
-
-  test('should throw error when presale configuration is missing', async () => {
-    const configWithoutPresale: ClankerTokenV4 = {
-      name: 'Test Token',
-      symbol: 'TEST',
-      tokenAdmin: '0x5b32C7635AFe825703dbd446E0b402B8a67a7051',
-      chainId: 8453,
-    };
-
-    await expect(clankerTokenV4PresaleConverter(configWithoutPresale)).rejects.toThrow(
-      'Presale configuration is required for presale converter'
-    );
-  });
-
-  test('should throw error when presaleEthToCreator is not available on chain', async () => {
-    // This will fail because presaleEthToCreator is not configured for Base chain
-    await expect(clankerTokenV4PresaleConverter(baseConfig)).rejects.toThrow(
-      'PresaleEthToCreator is not available on chain 8453'
-    );
-  });
-
-  test('should create correct transaction structure when presale is configured', async () => {
-    // We'll test the structure by catching the expected error but checking the transaction structure
-    try {
-      await clankerTokenV4PresaleConverter(baseConfig);
-    } catch (error) {
-      // We expect this error because presaleEthToCreator is not configured
-      expect((error as Error).message).toContain(
-        'PresaleEthToCreator is not available on chain 8453'
-      );
-    }
-  });
-
-  test('should handle presale configuration with default recipient', async () => {
-    const configWithDefaultRecipient: ClankerTokenV4 = {
-      ...baseConfig,
-      presale: {
-        minEthGoal: 1,
-        maxEthGoal: 10,
-        presaleDuration: 3600,
-        // recipient not specified - should default to tokenAdmin
-        lockupDuration: 0,
-        vestingDuration: 0,
-      },
-    };
-
-    try {
-      await clankerTokenV4PresaleConverter(configWithDefaultRecipient);
-    } catch (error) {
-      // We expect this error because presaleEthToCreator is not configured
-      expect((error as Error).message).toContain(
-        'PresaleEthToCreator is not available on chain 8453'
-      );
-    }
-  });
-
-  test('should handle presale with other extensions', async () => {
-    const configWithExtensions: ClankerTokenV4 = {
-      ...baseConfig,
-      vault: {
-        percentage: 20,
-        lockupDuration: 604800, // 7 days minimum
-        vestingDuration: 86400,
-      },
-      airdrop: {
-        merkleRoot: '0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef',
-        lockupDuration: 86400,
-        amount: 250000000, // Minimum amount
-      },
-      devBuy: {
-        ethAmount: 0.1,
-        poolKey: {
-          currency0: '0x0000000000000000000000000000000000000000',
-          currency1: '0x0000000000000000000000000000000000000000',
-          fee: 0,
-          tickSpacing: 0,
-          hooks: '0x0000000000000000000000000000000000000000',
-        },
-        amountOutMin: 0,
-      },
-    };
-
-    try {
-      await clankerTokenV4PresaleConverter(configWithExtensions);
-    } catch (error) {
-      // We expect this error because presaleEthToCreator is not configured
-      expect((error as Error).message).toContain(
-        'PresaleEthToCreator is not available on chain 8453'
-      );
-    }
-  });
-
-  test('should handle presale with custom fees', async () => {
-    const configWithCustomFees: ClankerTokenV4 = {
-      ...baseConfig,
-      fees: {
-        type: 'dynamic',
-        baseFee: 100,
-        maxFee: 500,
-        referenceTickFilterPeriod: 300,
-        resetPeriod: 3600,
-        resetTickFilter: 100,
-        feeControlNumerator: 1000,
-        decayFilterBps: 9500,
-      },
-    };
-
-    try {
-      await clankerTokenV4PresaleConverter(configWithCustomFees);
-    } catch (error) {
-      // We expect this error because presaleEthToCreator is not configured
-      expect((error as Error).message).toContain(
-        'PresaleEthToCreator is not available on chain 8453'
-      );
-    }
-  });
-
-  test('should handle presale with vanity address', async () => {
-    const configWithVanity: ClankerTokenV4 = {
-      ...baseConfig,
-      vanity: true,
-    };
-
-    try {
-      await clankerTokenV4PresaleConverter(configWithVanity);
-    } catch (error) {
-      // We expect this error because presaleEthToCreator is not configured
-      expect((error as Error).message).toContain(
-        'PresaleEthToCreator is not available on chain 8453'
-      );
-    }
-  });
-
-  test('should handle presale with custom pool configuration', async () => {
-    const configWithCustomPool: ClankerTokenV4 = {
-      ...baseConfig,
-      pool: {
-        pairedToken: '0x1234567890123456789012345678901234567890',
-        tickIfToken0IsClanker: -100000,
-        tickSpacing: 100,
-        positions: [
-          {
-            tickLower: -100000,
-            tickUpper: -99000,
-            positionBps: 10000,
-          },
-        ],
-      },
-    };
-
-    try {
-      await clankerTokenV4PresaleConverter(configWithCustomPool);
-    } catch (error) {
-      // We expect this error because presaleEthToCreator is not configured
-      expect((error as Error).message).toContain(
-        'PresaleEthToCreator is not available on chain 8453'
-      );
-    }
-  });
-
-  test('should handle presale with custom rewards configuration', async () => {
-    const configWithCustomRewards: ClankerTokenV4 = {
-      ...baseConfig,
-      rewards: {
-        recipients: [
-          {
-            admin: '0x5b32C7635AFe825703dbd446E0b402B8a67a7051',
-            recipient: '0x5b32C7635AFe825703dbd446E0b402B8a67a7051',
-            bps: 7000,
-            token: 'Both',
-          },
-          {
-            admin: '0x1234567890123456789012345678901234567890',
-            recipient: '0x1234567890123456789012345678901234567890',
-            bps: 3000,
-            token: 'Clanker',
-          },
-        ],
-      },
-    };
-
-    try {
-      await clankerTokenV4PresaleConverter(configWithCustomRewards);
-    } catch (error) {
-      // We expect this error because presaleEthToCreator is not configured
-      expect((error as Error).message).toContain(
-        'PresaleEthToCreator is not available on chain 8453'
-      );
-    }
-  });
-
-  test('should handle presale with sniper fees', async () => {
-    const configWithSniperFees: ClankerTokenV4 = {
-      ...baseConfig,
-      sniperFees: {
-        startingFee: 500000,
-        endingFee: 50000,
-        secondsToDecay: 30,
-      },
-    };
-
-    try {
-      await clankerTokenV4PresaleConverter(configWithSniperFees);
-    } catch (error) {
-      // We expect this error because presaleEthToCreator is not configured
-      expect((error as Error).message).toContain(
-        'PresaleEthToCreator is not available on chain 8453'
-      );
-    }
-  });
-});
